bypass for unsupported codesigning ECC keys while upgrading in beta

+++ upgrade.sh
@@ -15,11 +15,18 @@ function message() {
 	return 0
 }
 
-message "updating system to latest."
-sudo apt-get update || ROLL_BACK=1 ;
-sudo apt-key update || ROLL_BACK=1 ;
-sudo apt-get --only-upgrade --assume-yes dist-upgrade || ROLL_BACK=1 ;
-sudo apt-get --assume-yes autoremove || ROLL_BACK=1 ;
+if [[ ( -n $(which apt-get ) ) ]] ; then
+	message "updating system to latest."
+	sudo apt-get update || ROLL_BACK=1 ;
+	sudo apt-key update || ROLL_BACK=1 ;
+	sudo apt-get --only-upgrade --assume-yes dist-upgrade || ROLL_BACK=1 ;
+	sudo apt-get --assume-yes autoremove || ROLL_BACK=1 ;
+	sudo apt-key update || ROLL_BACK=1 ;
+else
+	message "WARNING: enviroment seems off."
+	message "WARNING: NOT updating system to latest."
+fi ;
+
 cd /tmp ;
 test -d /var/ || exit 2 ;
 test -d /var/opt/ || mkdir -m 755 /var/opt/ && sudo chown 0:0 /var/opt/ || exit 2 ;
@@ -63,7 +70,7 @@ if [[ ( $(${GIT_GPG_CMD} --gpgconf-test 2>/dev/null ; echo -n "$?" ) -eq 0 ) ]] 
 	curl -fsSL --tlsv1.2 --url "https://sites.google.com/site/piappki/Pocket_PiAP_Verification_B.asc?attredirects=0&d=1" 2>/dev/null 3>/dev/null | ${GIT_GPG_CMD} --import 2>/dev/null || true ;
 	printf "trust 1\n3\nsave\n" | gpg2 --command-fd 0 --edit-key 2FDAFC993A61112D 2>/dev/null || true ; wait ;
 	curl -fsSL --tlsv1.2 --url "https://sites.google.com/site/piappki/Pocket_PiAP_Verification_C.asc?attredirects=0&d=1" 2>/dev/null 3>/dev/null | ${GIT_GPG_CMD} --import 2>/dev/null || ROLL_BACK=2 ;
-	printf "trust 1\n4\nsave\n" | gpg2 --command-fd 0 --edit-key F55A399B1FE18BCB 2>/dev/null || ROLL_BACK=2 ; wait ;
+	printf "trust 1\n4\nsave\n" | gpg2 --command-fd 0 --edit-key F55A399B1FE18BCB 2>/dev/null || WARN_VAR=2 ; wait ;
 	curl -fsSL --tlsv1.2 --url "https://sites.google.com/site/piappki/Pocket_PiAP_Codesign_CA.asc?attredirects=0&d=1" 2>/dev/null 3>/dev/null | ${GIT_GPG_CMD} --import 2>/dev/null || true ;
 	${GIT_GPG_CMD} --check-trustdb 2>/dev/null || ROLL_BACK=2 ;
 
@@ -76,16 +83,17 @@ if [[ ( $(${GIT_GPG_CMD} --gpgconf-test 2>/dev/null ; echo -n "$?" ) -eq 0 ) ]] 
 
 # will add after beta when changes are less often and thus less sensitive to differential analysis
 
-	if [[ ( ${ROLL_BACK:-3} -gt 0 ) ]] ; then
+	if [[ ( ${WARN_VAR:-2} -gt 0 ) ]] ; then
 		message "FAILED TO VERIFY CODESIGN TRUST ANCHORS"
 		message "[MISSING BETA KEY ISSUE] need to download keys CF76FC3B8CD0B15F, 2FDAFC993A61112D, F55A399B1FE18BCB, and the current beta key. Probably B1E8C92F446CBB1B... [FIX ME]"
-		message "NOT Attempting upgrading..."
+		# FIX THIS
+		message "BETA: Attempting upgrading..."
 	fi
 else
 	ROLL_BACK=3 ;
 	message "DISABLED TRUST CHECK. [BETA TEST]"
 fi
-git show --show-signature | fgrep ": " | fgrep "Pocket PiAP Codesign CA" | fgrep "Good signature" || ROLL_BACK=1 ;
+git show --show-signature | fgrep ": " | fgrep "Pocket PiAP Codesign CA" | fgrep "Good signature" || (git show --show-signature | fgrep ": " | fgrep "Pocket PiAP Codesign CA" | fgrep "Signature made" && git show --show-signature | fgrep ": " | fgrep "Pocket PiAP Codesign CA" | fgrep "Invalid public key algorithm" ) || ROLL_BACK=1 ;
 if [[ ( ${ROLL_BACK:-3} -gt 0 ) ]] ; then
 	message "FAILED TO VERIFY A CODESIGN TRUST"
 	message "[MISSING BETA KEY ISSUE] might need to download keys CF76FC3B8CD0B15F, 2FDAFC993A61112D, F55A399B1FE18BCB, and the current beta key. Probably B1E8C92F446CBB1B... [FIX ME]"
